utils/testcase.bash: backup only existing files/dirs
Fix backup function to skip not existing files/dirs.
This fix is backward compatible and has no effect
on existing usage of the function.

+++ audit-test/utils/testcase.bash
@@ -72,8 +72,11 @@ function append_cleanup {
 function backup {
     declare f b
     for f in "$@"; do
-	b=$(mktemp "$f.XXXXXX") || exit_error
-	cp -a "$f" "$b" || exit_error
-	prepend_cleanup "mv -f '$b' '$f'"
+    # if the path/file does not exist - skip backup
+    if [ -e "$f" ]; then
+        b=$(mktemp "$f.XXXXXX") || exit_error
+        cp -a "$f" "$b" || exit_error
+        prepend_cleanup "mv -f '$b' '$f'"
+    fi
     done
 }
